feat: изменена архитектура
