--- src/Components/Common/index.js
@@ -1,1 +1,0 @@
-export * from "./LoadingSpinner";
