--- src/Components/Pages/CharactersPage.jsx
@@ -1,8 +1,0 @@
-import CharsterCard from "../CharsterCard";
-import { CommonPage } from "../Common/CommonPage";
-
-const CharactersPage = () => {
-  return <CommonPage CardComponent={CharsterCard} />;
-};
-
-export default CharactersPage;
