+++ src/components/commonPage/index.js
@@ -1,0 +1,1 @@
+export * from "./CommonPage.jsx";
