+++ src/components/index.js
@@ -1,0 +1,4 @@
+export * from "./commonPage";
+export * from "./loadingSpinner";
+export * from "./logo";
+export * from "./privatRout";
