+++ src/components/privatRout/index.js
@@ -1,0 +1,1 @@
+export * from "./PrivatRoute.jsx";
