+++ src/layouts/OneElemPageLayout/index.js
@@ -1,0 +1,1 @@
+export * from "./OneElemPageLayout.jsx";
