+++ src/layouts/index.js
@@ -1,0 +1,2 @@
+export * from "./LocationsAndEpisodesPageLayout";
+export * from "./OneElemPageLayout";
