+++ src/main.jsx
@@ -1,6 +1,6 @@
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import App from "./App.jsx";
+import { App } from "./app";
 import "./index.css";
 
 createRoot(document.getElementById("root")).render(
