+++ src/pages/charactersPage/CharactersPage.jsx
@@ -1,0 +1,8 @@
+import CharsterCard from "./components/characterCard/CharsterCard";
+import { CommonPage } from "../../components";
+
+const CharactersPage = () => {
+  return <CommonPage CardComponent={CharsterCard} />;
+};
+
+export default CharactersPage;
