+++ src/pages/homePage/components/HeaderCustom/components/index.js
@@ -1,0 +1,2 @@
+export * from "./navbar/Navbar.jsx";
+export * from "./sortingField/SortingField.jsx";
