+++ src/pages/homePage/components/HeaderCustom/components/sortingField/SortingField.jsx
@@ -1,0 +1,23 @@
+import { useSearchParams } from "react-router-dom";
+import { Select } from "antd";
+
+export const SortingField = () => {
+  const [searchParams, setSearchParams] = useSearchParams();
+
+  const handleChange = (e) => {
+    setSearchParams({ sortByCreate: e });
+  };
+
+  return (
+    <Select
+      defaultValue="Oldest to Newest"
+      size={window.innerWidth < 768 ? "small" : "large"}
+      style={{ width: 150 }}
+      onChange={handleChange}
+      options={[
+        { value: "createdASC", label: "Oldest to Newest" },
+        { value: "createdDESC", label: "Newest to Oldest" },
+      ]}
+    />
+  );
+};
