+++ src/pages/homePage/components/HeaderCustom/index.js
@@ -1,0 +1,1 @@
+export * from "./HeaderCustom.jsx";
